+++ src/cli.ts
@@ -1,4 +1,4 @@
-import {DaemonClient} from '#daemon/DaemonClient'
+import {daemonClient} from '#daemon/DaemonClient'
 import {loadAndAssertConfig} from '#lib/validateConfig'
 import {createReadStream, watchFile} from 'fs'
 import {join} from 'path'
@@ -9,46 +9,46 @@ const command = process.argv[2]
 switch (command) {
   case 'start': {
     loadAndAssertConfig(join(import.meta.dirname, '../conf.d'))
-    if (await DaemonClient.isRunning()) {
+    if (await daemonClient().isRunning()) {
       console.log('Server is already running')
       process.exit(0)
     }
     process.stdout.write('Starting server... ')
-    await DaemonClient.start()
+    await daemonClient().start()
     console.log('done')
     break
   }
 
   case 'stop': {
-    if (!await DaemonClient.isRunning()) {
+    if (!await daemonClient().isRunning()) {
       console.log('Server is not running')
       process.exit(0)
     }
     process.stdout.write('Stopping server... ')
-    await DaemonClient.stop()
+    await daemonClient().stop()
     console.log('done')
     break
   }
 
   case 'restart': {
     loadAndAssertConfig(join(import.meta.dirname, '../conf.d'))
-    if (await DaemonClient.isRunning()) {
+    if (await daemonClient().isRunning()) {
       process.stdout.write('Stopping server... ')
-      await DaemonClient.stop()
+      await daemonClient().stop()
       console.log('done')
     }
     process.stdout.write('Starting server... ')
-    await DaemonClient.start()
+    await daemonClient().start()
     console.log('done')
     break
   }
 
   case 'status': {
-    if (!await DaemonClient.isRunning()) {
+    if (!await daemonClient().isRunning()) {
       console.log('Server: stopped')
       process.exit(0)
     }
-    const response = await new DaemonClient().send({type: 'status'})
+    const response = await daemonClient().send({type: 'status'})
     if (response.type !== 'status') {
       console.error('Unexpected response from daemon')
       process.exit(1)
@@ -72,7 +72,9 @@ switch (command) {
     const readFrom = (start: number) => {
       const s = createReadStream(logFile, {encoding: 'utf8', start})
       s.pipe(process.stdout, {end: false})
-      s.on('end', () => { pos += s.bytesRead })
+      s.on('end', () => {
+        pos += s.bytesRead
+      })
     }
     readFrom(0)
     watchFile(logFile, {interval: 300}, () => readFrom(pos))